move bubble in atoms
