+++ client/src/components/_atoms/DogsBubbles/DogsBubbles.tsx
