+++ client/src/components/_atoms/TrainerBubble/TrainerBubble.scss
