+++ client/src/components/_molecules/ParticipantsOverview/ParticipantsOverview.scss
